pyproject.toml
[project]
name = "s7-mcp-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "fastmcp>=2.4.0",
    "httpx>=0.28.1",
    "icecream>=2.1.4",
    "pydantic>=2.11.5",
    "smolagents[mcp,openai]>=1.16.1",
]

[dependency-groups]
dev = [
    "marimo>=0.13.11",
    "ruff>=0.11.11",
    "ty>=0.0.1a6",
]
